feat: Change build tool (from Premake) to CMake
Previous project generation tool was Premake. It is changed to favor compatibility with various open source projects, as they commonly make use of CMake as their project generation tool. This was a long postponed tech debt that had to be paid before project gets any more complicated than this.

Apart from this, simple clean up is done to get rid of all premake remainings as well as renamed place-holder files to .gitkeep to be more universal in naming convention.